refactor admin store into slices
